minor bug fix

+++ src/rl_agents/environments/envs/localize_env.py
@@ -146,7 +146,7 @@ class LocalizeGibsonEnv(iGibsonEnv):
 
         self.pf_params.return_state = True
         self.pf_params.stateful = False
-        self.pf_params.global_map_size = np.array(FLAGS.global_map_size, dtype=np.float32)
+        self.pf_params.global_map_size = np.array(FLAGS.global_map_size, dtype=np.int32)
         self.pf_params.window_scaler = FLAGS.window_scaler
         self.pf_params.max_step = self.config.get('max_step', 500)
